Remove last two js files, fix tsconfig to output dist without doubling up dirs
